transformDate method moved to Date Helper
